Add usage statement

+++ deploy_functions.sh
@@ -1,24 +1,21 @@
 #!/bin/sh
 
-# Packages and deploys Lambda Functions
-mkdir -p packages
-
 NODEJS_VERSION="nodejs14.x"
 PYTHON_VERSION="python3.9"
 RUBY_VERSION="ruby2.7"
 JAVA_VERSION="java11"
-
 DEFAULT_FUNCTION_SIZE=1024
-FUNCTION_SIZE=${1:-$DEFAULT_FUNCTION_SIZE}
 FUNCTION_SUFFIX="hello-world"
 POLICY_DOCUMENT=`cat assume-role-policy-document.json`
 
 create_lambda() {
 	aws lambda get-function --function-name "$FUNCTION_SIZE-$1-$FUNCTION_SUFFIX" > /dev/null 2>&1
 	if [ $? -eq 0 ]; then
+		echo "$FUNCTION_SIZE-$1-$FUNCTION_SUFFIX exists already."
 		echo "Updating $1 Lambda Function..."
 		aws lambda update-function-code --function-name "$FUNCTION_SIZE-$1-$FUNCTION_SUFFIX" --zip-file "fileb://packages/$1.zip"
 	else
+		echo "$FUNCTION_SIZE-$1-$FUNCTION_SUFFIX does not exist."
 		echo "Creating $1 Lambda Function..."
 		
 		aws iam get-role --role-name "$FUNCTION_SUFFIX" > /dev/null 2>&1
@@ -34,19 +31,39 @@ create_lambda() {
 	fi
 }
 
+
+if echo $* | grep -e "help\|man\|usage" -q; then
+	echo "Usage:"
+	echo "sh ./deploy_functions.sh [function_size=1024] [create_packages]\n"
+	echo "Options:"
+	echo "function_size		Lambda Function Memory (MB)"
+	echo "create_packages		Compile and create ZIP packages"
+	echo "\n\n"
+	exit 0
+fi
+	
+
+if echo $* | grep -e "create_packages" -q; then
+	echo "Creating packages..."
+	mkdir -p packages
+	cd nodejs && zip ../packages/nodejs.zip index.js && cd ..
+	cd python && zip ../packages/python.zip lambda_function.py && cd ..
+	cd ruby && zip ../packages/ruby.zip lambda_function.rb && cd ..
+	cd java && gradle buildZip && cd .. && cp java/build/distributions/example.zip packages/java.zip
+fi
+
+MAYBE_FUNCTION_SIZE=`echo $* | grep "function_size" | cut -d "=" -f2`
+FUNCTION_SIZE=${MAYBE_FUNCTION_SIZE:-$DEFAULT_FUNCTION_SIZE}
+
 ## Package and deploy nodeJS
-cd nodejs && zip ../packages/nodejs.zip index.js && cd ..
 create_lambda nodejs "$NODEJS_VERSION" "index.handler"
 
 ## Package and deploy Python
-cd python && zip ../packages/python.zip lambda_function.py && cd ..
 create_lambda python "$PYTHON_VERSION" "lambda_function.lambda_handler"
 
 ## Package and deploy Ruby
-cd ruby && zip ../packages/ruby.zip lambda_function.rb && cd ..
 create_lambda ruby "$RUBY_VERSION" "lambda_function.lambda_handler"
 
 ## Package and deploy Java
-cd java && gradle buildZip && cd .. && cp java/build/distributions/example.zip packages/java.zip
 create_lambda java "$JAVA_VERSION" "example.Hello::handleRequest"
